// frontend/src/lib/api.ts
export interface MinistryTotal {
  ministry: string;
  total: number;
  divisions: { division: string; total: number }[];
}